+++ Formula/stelekit.rb
@@ -2,16 +2,16 @@
 class Stelekit < Formula
   desc "Markdown-based outliner and note-taking app (Kotlin Multiplatform)"
   homepage "https://github.com/tstapler/stelekit"
-  version "0.41.3"
+  version "0.45.0"
 
   on_linux do
     url "https://github.com/tstapler/stelekit/releases/download/v#{version}/SteleKit-v#{version}-linux.AppImage"
-    sha256 "272ea5ce01e9dff8341245d34f6b65391901022d123f9e7dd51d2e6755452b41"
+    sha256 "3e83ab0c01cc175002c24492a6721e71c8b7e21754538d6cab36a6fd2ee4a880"
   end
 
   on_macos do
     url "https://github.com/tstapler/stelekit/releases/download/v#{version}/SteleKit-v#{version}-macos.dmg"
-    sha256 "98f4bcc4367372c44d63abf80aa9201cb7c2acc6d8e1f4d718be36e56dd4c5ce"
+    sha256 "95866b62dba34db6ed2158e7e543d476345accbc7996d6612bbb6cd76b7dac78"
 
     depends_on "openjdk"
   end